v1.0.2

+++ internal/decoders/addressing.go
@@ -112,7 +112,7 @@ func decodeAddressingMode(data []byte, mode, reg uint8, operandSize int) (string
 				return "", 0, Operand{}, err
 			}
 			addr := int16(binary.BigEndian.Uint16(data[:2]))
-			absolute := uint32(uint16(addr))
+			absolute := uint32(int32(addr))
 			text := fmt.Sprintf("$%04X", uint16(addr))
 			return text, 1, effectiveAddressOperand(text, EffectiveAddress{
 				Kind:            EAKindAbsoluteShort,
